Serverless/utils_python/useful/zzz_cdk_apigwlambdadynamopoetry/requirements.txt
# AWS CDK Dependencies
aws-cdk-lib
constructs
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha
aws-cdk.aws-lambda-python-alpha

# Lambda Dependencies
boto3==1.28.62
aws-lambda-powertools==2.26.0
aws_xray_sdk==2.14.0

# Development Dependencies
pytest
pytest-mock
pytest-pythonpath